Save all values in self.values
This way, we can directly return self.values when getting the preferences from QML. No need to make a copy any more.
A copy is made when loading the preferences, but that really only happens once. A copy is also made when saving the preferences, but changing a preference is much less frequent than getting a preference. So I hope this improves performance a little bit.