Fix getting fields by name from database
Otherwise it escapes the field name, causing it to return that field name as a literal string.
Right now the field name would be sensitive to SQL injection, so we need to be sure that no user-defined input can determine the field name to get here. For now that's impossible, luckily.