Fix blending shadereffect with background
So it turns out that the shader effect also shades the background behind the image, not just the image as a separate fully transparent buffer. I'm still not 100% clear on why this happens or how exactly it plays out. However I did manage to create a fix by manually alpha-blending it.