Add global storage of metadata
I aim to keep the single source of truth in control in Python rather than inside of the database. You'd think that the database would be very fast at serving data, but it turns out to not be very fast at that and much more complex due to threading. It may be good at processing complex queries, but really all we need is a table. On my own computers I'll probably display 95% of the music in a table anyway, and I'll always look up which songs to display by file path. So this dictionary is fine.
The dictionary will need to be filled from the database file in a deserialisation and serialisation step. That still needs to be implemented then.