Store metadata as dicts instead of namedtuples
Dicts are more efficient in general, but the difference with a named tuple is so minimal that it wouldn't matter. However now that I'm re-implementing the music directory, I'm finding that I have to convert everything to dict there. So instead, I'll just let this return dicts and then there is no conversion step necessary which is easier to program, faster and more elegant in general.