Re-implement MusicDirectory as a table model
That makes it a lot easier to sort the table.
The sorting is currently not persistent if the directory changes. That'll need to be added later once I've got this to work properly.
Currently this table doesn't work. It's getting errors in there somewhere. But I do want to commit this just so that I have a baseline to work from while I'm experimenting to get it to work.