Fix priority direction of sorting list
The sorting function was giving the first element the highest sorting priority due to the nature of how the keys are generated.
In order to keep the key generation simple and fast we'd like to keep it that way. But then the ordering of the sort_field list has to be inverted.