Fix QML warnings on start-up about zero-size implicit size
The width can just be a fixed size. It will get changed when it re-computes the layout before the first render anyway.
The height needs to be the height of the text. But I think for some elements this height ends up at 0, perhaps for cells without text. So give it some height then to make Qt shut up about it.