Add function to change an element of a list or dict
Otherwise it's a bit hard to cause it to register that there's been a change.
Although this function documents that it works for both lists and dicts, and the Python code would do that, the PySide binding currently only allows the index argument to be integer. I might want to fix that in the future if I end up having a dict in the preferences somewhere. But currently I'm only using it for a list. I guess it could work if your dict keys are integers...