Add storage location for data files
We'll move our database there because it's going to contain more than just cached information from the metadata of the files. It will also need to contain data such as when the file was last played and how often. It'll need to contain waypoints for each file's volume control. And perhaps a link to a pre-generated spectral analysis image. This can't all be stored in the ID3 or FLAC metadata. And so the database will not just be a cache but also actually containing new data.