Allow sorting to be called from QML
This means that the function has to accept two types of input data for the column field, an integer (for when it's called by Qt) or a string (for when it's called from QML with a role name).
I don't know how to trigger this from being called by Qt the proper way. Perhaps it's something that the table headers could normally do, but it's not documented.