Make it sort descending when clicking column twice
When a column is the highest priority sort and is then clicked again, reverse the sorting order.
To implement this we've introduced a function that can tell whether a column is the most recently sorted column and what the sort direction was. We'll need this to display an arrow in the column header to indicate the sorted column.