Implement arrow key navigation
Originally I also had tab navigation do the same thing. However Qt already captures the tab key to navigate between different elements, so it didn't work to use the tab key if there is an element after it in the flow, and it didn't work to use shift+tab if there is an element before it in the flow that can have focus. The arrow keys will have to do.