Keep selected row in view (when it changes)
This way, when you use the arrow keys to navigate the list, the view of the table will scroll along. Also in the future, if selecting an item from the playlist, it will select the same track in the music table and so it will then scroll to that row too.
This attempts to mimic the ListView.positionViewAtIndex(row, ListView.Visible) function.
This only works if all of the items in the table have the same height, which is currently the case. I hope it remains the case.