Re-implement drag and drop for playlist items
The old approach was to re-order elements of the list during the dragging. Items would then automatically move according to their new position in the list, even leaving a space for the item that was being dragged.
However that didn't work, because the data in the list was being swapped, even the data on the item being dragged. So you were then suddenly dragging a different track.
This new approach manually re-positions elements during the dragging but doesn't actually swap them in the list until the mouse is released. And it seems to actually work.