Fix dragging items back without releasing them
Previously it was re-positioning every element between the old position and the new position. If you were to drag an item quite far, but then without releasing the mouse drag it back, then the farthest few elements would have been re-arranged but are never considered for being put back when you drag back.
Now it'll re-arrange all items. Some of them multiple times. It will first re-position every item to their proper position and then re-consider for the range between old_index and new_index what the modified position should be. This way the items outside of that range will be placed back in their proper place.
Not very efficient, but this list should really just contain like 6 items most of the time, so then it's no problem.