When selecting from playlist, also select in file browser
This is a way in which you can find files back in the file browser from what's in the playlist. The playlist doesn't display all information and some of it is not displayed accurately (BPM is only by the background colour, not numeric). So to find that additional information you can now select it in the playlist and look up in the table what the detailed metadata is. Very handy.