Select item when dragging
This prevents the slightly odd behaviour of re-selecting the current item in the file browser when dragging other items around it. Previously it would re-select the item in the file browser when dragging other playlist items such that the position of the selected item changes in the list. Now it will immediately select the dragged item and that appears more logical to the user.