Whitelist which fields you can edit with change dialogue
You can't edit the duration of a song with a text field. And in the future we'll want to add new fields to this table such as when it was last played. Some of those fields may be editable as well, but some will not be. That's why there's a whitelist here.