Store playlist in preferences
This required a bit of extra access to the internals of the preferences, especially since we're also storing dicts inside of lists and editing that dict. But it works pretty much out of the box with that.
I don't like how it stores the entire dict in the preferences now. But it's fine. It's easy to work with this way. Just doesn't keep things in the playlist in sync if the metadata changes. But the title of the song is the only relevant one really that could ever be changed, and I never really change the title. If the title gets outdated in the playlist, it's not the end of the world.