Choose Pygame as media player to play sound files
This is a well-considered choice. I've looked into the following constraints for the audio media player:
- It needs to support Linux and Windows.
- It needs to be easy to install via Pip.
- It needs to be capable of playing MP3 and FLAC files at minimum.
- It needs to be able to pause and resume playback.
- It needs to have responsive volume control (in order to implement fade-outs when pausing the music).
- It needs to have an equaliser of some sort that can adjust the music EQ on the fly during playback.
- It needs to be supported and recently maintained.

The constraint for it to have an equaliser seems to be impossible to fulfil. There are no Python packages that I could find that play audio files with an equaliser. I would prefer to use Qt, which matches all of the other constraints but doesn't have an equaliser. It might be possible to get Qt to play from an audio source created by some convoluted way, but this would be very hard.
Instead, I chose Pygame. It also doesn't have an equaliser built-in, but it does expose its audio frames to the Python environment, from where the waveform can be edited. I've seen examples where people have used Scipy to then apply an EQ filter to the audio via a FFT and inverse FFT. This is exactly the sort of thing I need. I hope it'll sound good, because I haven't tried it yet. Let's see.