Add function to unload a track from memory
I considered automatically unloading it when removing a track from the playlist (in its remove function). However I'm imagining that the remove function will also be called by the player in order to remove the track that is currently playing. That's the precise moment when you'd probably NOT want the audio to unload... So I can't unload it then.
This probably means that there's going to be a memory leak: An audio file will get loaded when you add it to the queue and will remain loaded even if you remove it from the queue with the unqueue button. I can make it unload when the file has been played, but won't make it unload if the user removes it manually from the queue.
It's not too bad though. It shouldn't occur frequently and if it does, my computer has plenty of memory to keep all of that in RAM, honestly.