Turn Playlist into a singleton
I've tried various approaches to get to a playlist instance from within the player Python class. But it's just very weird with passing that on through QML. I've encountered several bugs in Pyside that I was unable to find solutions to online. This is the simplest way to do it.