Don't pre-load files
It turns out that during the pre-loading, the music temporarily stops, as if the pre-loading takes up the GIL and that also prevents the music itself from playing. It's not a GIL of course, but something similar seems to be going on inside Pygame. It's weird.
Instead of pre-loading, it'll have to load the track when it starts playing. It adds a few fractions of a second to the start time, but nobody will really notice since there should be a few full seconds between tracks anyway. No gapless playback necessary.