Remove track from playlist when it's done playing
That way, if we stop the music, the last played track will remain in front of the queue to play again.
And since we don't have the title shown yet it also doubles as showing the title of the playing game. And it makes more sense with the cumulative time remaining thing if we get that to update during playing.