Turn Player into a singleton to ensure isPlaying signal arrives correctly
If you have multiple instances of this, and one starts playing, the other instance don't notice that. So their isPlaying property stays unchanged as far as QML is concerned. Even using a class-level signal doesn't work, since it gives error when you use a signal like that. So there really can only be one instance of the player.
Lots of singletons in this application. But it doesn't really matter. It's not likely to grow out of hand.