Show fourier image in interface
Looks very ugly right now. There'll be more spacing on the left for the volume and EQ sliders at some point. And there will be more spacing at the top and bottom for the title and timestamps. But for now this is fine to see how to make this fit in with the colour scheme of the interface (which is not done yet at the moment).