Add a style metadata field
This field is intended for a categorisation of the sort of band or the sort of music it is, such as big band, stompy, vocal-oriented, sensitive, etc. It's again a free-form field so it's up to the user to be consistent with it. But if it is used consistently, it can be useful for sorting and finding certain types of songs.