Add database entries for the 4 types of waypoints
We'll want to store waypoints in the database for the volume, the bass volume, mids volume and treble volume eventually. Volume will be the easiest one to implement.
By adding all 4 of these to the database already, hopefully there won't be any more database changes for a while. Every time I make those database changes I'm erasing the database on my computer because I don't have an upgrade mechanism yet.