Fail gracefully if preferences are corrupt
Otherwise the application won't start any more.
This happened to me during a DJ session when the application crashed due to the audio driver going offline. Apparently it was just writing the preferences file, and as such the preferences were written to become a 0-byte empty file. This then prevented me from restarting the application because of a parsing error.