Re-generate waypoint graph when resizing image
This is important because the image always starts off with a width and height of 0 in the init. Only after initialising the object, will Qt change the width and the height. So during the init, it still has a width and height of 0 which results in a nonsensical image. We must re-generate the SVG after Qt sets the proper width (and also if the user resizes the window and such).