Add events for music volume control based on waypoints
This now executes the waypoints, completing the data loop to actually become useful. It will now automatically remember volume adjustments, and execute them again the next time the song plays.
Still a problem with it though is that existing events are created at the start of playback and will not be erased when new waypoints are added during playback. This makes the volume jump a bit sometimes. Got to find a way to deal with that.