Update theme when theme preference changes
This allows the entire interface to be updated on the fly whenever the theme changes. It affects pretty much every single QML element. Pretty impressive!
Incidentally, it takes Qt a second to change it fully. It seems that it's all pretty quickly updated in a few milliseconds, except the current playlist which has some transparent elements that take a little longer it seems.