Set QSG_RENDER_LOOP to basic to fix animation speed
Due to a Qt bug, the animation speed is inaccurate when using the threaded render loop of Qt. We're using the animation speed to display the current position in the currently playing track (which is arguably not a good idea, but oh well, it performs better than the only alternative I can think of). As such, the animation speed being inaccurate results in an inaccurate display of where in the song we currently are.
See also the related Qt bug: https://bugreports.qt.io/browse/QTBUG-58885 (linking here for posterity if someone is looking through all of Github's commits for that bug or something).