Add a cumulative endtime to the playlist model
This is a timestamp of when an item on the playlist is about to end.
This will cause the preferences to be updated every second, even if the music is not playing. Not very efficient, and that needs a remedy in the future. But for now I'll keep it as it is.