Only include history for the current session
We want the time it was last played to factor in as well. But to get a lot of variation in a DJ session you want to know which songs were played in this session. That information is obtained by looking at the history. The last played time to add to the score will just be obtained from the metadata again.