Decode audio files using pyminiaudio
We were decoding them using Pydub, but Pydub uses Ffmpeg or LibAV for decoding. Installing FFMpeg is simple enough, but packaging a binary with our application is a lot more involved. Instead of that, I'm removing the requirement on these applications by using a different library to decode audio files.
If FFMpeg or LibAV is not present, this will raise a warning. But it's just a warning.