Provide full platform URLs in theme icons
Because on some platforms, the simple path is not enough. On Linux just having /path/to/file.svg works, but on Windows, C:/path/to/file.svg causes an error because the C protocol is unknown. It needs to be file:///c/path/to/file.svg there. This does the proper conversion that Qt needs.