Add metadata to database if it's not in there yet
Because it could happen that a song is in the playlist, but is not yet in the database because the directory is pointed somewhere else. The user can't naturally get in this situation, but it's possible to get there with file editing (or just by deleting the database, which is what I did).