Track history just from current session
The previous way to track the history was problematic in a couple of ways. For one, it would arbitrarily depend on a 24 hour limit. If you were playing often (every day for instance) it would include some of the history from a previous session. Secondly, it used the preference item of a different class, which would make it depend on the initialisation order which class gets to create the preference. And thirdly, it depended on the browse path. Files that are not currently in the browse path would not get shown in the history. This would be fine if people work like me with 1 big folder for all music, but not if the music is split over different directories.
This new way has the disadvantage that if the application crashes or is reset the history is also reset. But let's assume that it won't crash, and why would the user restart the application? So I think this is better. It's also simpler.