Fix updating length of playlist with AutoDJ
It needs to update if the AutoDJ is added or removed, so it needs to be updated depending on the new list during the update() function. However the update() function is also called when something is manually added or removed, and so it would update the length twice, causing issues. So to fix that, we're no longer updating the length when adding or removing, just leaving all of it to the update() function. This is less efficient, but reliable with automatic as well as manual changes.