Add missing typing
This gets rid of the duck-typing idea that you should be able to provide any type of argument and it'll work on the behaviour of the type rather than the actual type. But it's more common these days.
The types have not yet been checked with Mypy itself. There might be errors.