Only allow a certain whitelist of system libs on POSIX
I'm getting errors on some computers if they are not using Wayland. They then use the OpenGL library packaged in the AppImage which refers to a system EGL library that doesn't have symbols for Wayland, but those symbols are refered to by the AppImage packaged libraries.
The list of libraries here is all of the libraries that currently get packaged into the AppImage by my computer, except for the libraries listed here, uncommented: https://github.com/AppImageCommunity/pkg2appimage/blob/master/excludelist
If the package is listed in that file but commented out with a reason, I've copied that reason over to my spec file.