Ensure local directory exists before configuring file logger
Because the file logger needs to place the log file somewhere. If it can't, we'd get an exception right there. The directory needs to exist.
However the ensure_exists() function itself also logs. It can't log then to the file in this case, but it will log with the default configuration. I guess it won't produce any log output though because the default configuration is to not otput anything below warning.