Install files into correct subdirectory
I was copying all files in the dist directory, which is currently just an uninstaller and the LynDJ directory.
This results in the installation folder containing a LynDJ subdirectory as well, often resulting in a double nested directory: C:\Program Files\LynDJ\LynDJ\lyndj.exe.
Instead, take the subdirectory directly.