Add an upgrader to check if config files are too modern
Before we actually do release our application we should really have something like this, even if it is primitive.
The user may run newer versions of our application that have changed the configuration files in a way that we can't understand them in this old version. We should really say something to the user when this happens, but to keep the changes minimal before the release we'll just have this basic implementation: If any file is too modern, we simply abort the start-up of the application. The user hopefully finds the log file which would help to see what's happening. Maybe in a future version we'll implement a small dialogue to see what's happening and perhaps let the user continue at their own risk.