Implement config-too-modern dialogue as splash screen
I had to use the splash screen, though it really isn't a splash screen. When using a normal QML window you'll need to run the QML engine, which wants a main window to be shown. I could show the dialogue then as the main window, but then changing that dialogue is hard. You could theoretically have an item as the root of the QML which can have multiple windows and switches the visibility of them. However then I'd also get into the problem of trying to create the main window without the theme or preferences being defined and such, so then I'd also need to make sure the main window is lazy loaded and stuff. Too complex for this super-hidden feature to warrant. I'll use the splash screen.
The splash screen is a QtWidget thing, which defines itself in a bit of a different way. To keep things simple and contained here in the upgrader, I made a faux draw function that draws faux buttons and a simple click event handler that checks the coordinates of the click. Nothing special and not super neat, but it works well.