Store sound as numpy arrays instead of bytes
We'll convert it to bytes for playback at the end, but in intermediate storage it'll be stored as numpy arrays. This makes it easier to filter these sounds. We'll pretty much always filter these sounds, so this makes it more efficient since we don't have to keep converting it back and forth. Just looking at the functions we're changing in this commit makes it apparent that this simplifies the code a lot and makes it more efficient too.