Recalculate volume change events when changing waypoints
Previously the actual events kept on running. So there was no way to alter the future. When you slid the volume slider, the volume would immediately change, but future volume changes were still applied in the gradients, causing the volume to suddenly go up temporarily.
Now we'll re-calculate the volume changes with the new graph, causing the gradients to be re-calculated properly when one of the gradients changes.