Start playing volume waypoints from cut_start
Otherwise, with large cut_starts, the timers will start too late.
In order to prevent skipping volume changes before the cut_start, we'll also add an extra volume change at t=0 to the current volume. This one is always added, which was previously seen as a waste (since the volume always started at 0.5 anyway) but is now necessary. It's just one extra timer, so it'll be fine.