Change start time and current position before current song
It's a bit obtuse and will be hard to debug later on because the order of signals is important here. If the current_position is still on the old position, the song_changed signal will trigger a recalculation of the waypoints starting from the old current_position, which triggers the volume changes and stuff at the wrong time, usually skipping all or most of them.